Fix null bug in returned Object[] during Places.callAll
This bug would cause null values to be returned at the end of the
array if the number of places was not evenly divisible by the
number of mNodes. Because the last mNode is responsible for any
remainder in the places matrix, it may return an irregularly sized
array. For simplicity, we can just check the size of that final
returned array to ensure it is completely copied into the return
values array that is provided to the client.